Move RangesListFactory from internal to root package
